Fix MSVC compiler warnings across libfaac and frontend
- Explicitly cast integer and floating point conversions in
  channels.c, frame.c, quantize.c, and stereo.c to eliminate
  MSVC C4244 and C4305 data truncation warnings.
- Map POSIX strdup to _strdup under MSVC in charset.h to fix
  C4996 deprecation warnings.
- Zero-initialize struct WAVEFORMATEXTENSIBLE wave in input.c to fix
  C4701 uninitialized variable warning.

+++ libfaac/quantize.c
@@ -459,7 +459,7 @@ void CalcBW(unsigned *bw, int rate, SR_INFO *sr, AACQuantCfg *aacquantCfg,
     sfbOffsetLong[i] = l;
     aacquantCfg->max_cbl = i;
     aacquantCfg->max_l = l;
-    *bw = (float)l * rate / (BLOCK_LEN_LONG << 1);
+    *bw = (unsigned int)((float)l * (float)rate / (float)(BLOCK_LEN_LONG << 1));
 }
 
 // short-window grouping: keep spectrally-similar windows together so they
